модификаторы cch перенесены в src/metadata/chartscharacteristics
