+++ config/webpack.config.dev.js
@@ -208,7 +208,7 @@ module.exports = {
             // it's runtime that would otherwise processed through "file" loader.
             // Also exclude `html` and `json` extensions so they get processed
             // by webpacks internal loaders.
-            exclude: [/\.js$/, /\.html$/, /\.json$/, /\.md$/],
+            exclude: [/\.js$/, /\.mjs$/, /\.jsx$/, /\.html$/, /\.json$/, /\.md$/],
             loader: require.resolve('file-loader'),
             options: {
               name: 'static/media/[name].[hash:8].[ext]',
