+++ src/metadata/chartscharacteristics/index.mjs
@@ -1,12 +1,12 @@
 // модификаторы планов видов характеристик
 
 // предопределенные элементы
-//import cch_predefined_elmnts from "./cch_predefined_elmnts";
+import cch_predefined_elmnts from "./cch_predefined_elmnts";
 
 // дополнительные реквизиты и сведения
-//import cch_properties from "./cch_properties";
+import cch_properties from "./cch_properties";
 
 export default function ($p) {
-	//cch_predefined_elmnts($p);
-	//cch_properties($p);
+	cch_predefined_elmnts($p);
+	cch_properties($p);
 }
