+++ src/metadata/modifiers.js
@@ -7,7 +7,7 @@
 // import documents from './documents';
 //
 // // модификаторы планов видов характеристик
-// import chartscharacteristics from './chartscharacteristics';
+import chartscharacteristics from './chartscharacteristics';
 //
 // // модификаторы отчетов
 import reports from './reports';
@@ -19,7 +19,7 @@ import common from './common';
 export default function ($p) {
   // catalogs($p);
   // documents($p);
-  // chartscharacteristics($p);
+  chartscharacteristics($p);
   reports($p);
   common($p);
 }
